+++ .travis.yml
@@ -9,6 +9,8 @@ services:
 
 before_script:
   - psql -c 'create database travis_ci_test;' -U postgres
+
+  - npm run mig-test
  
 script:
  - npm test
