+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "cross-env NODE_ENV=test  nyc mocha ./server/test/*.spec.js --require @babel/register --exit",
     "start:dev": " nodemon  app.js --exec babel-node",
-    "start": "baapp.js",
+    "start": "babel-node app.js --exec  babel-node",
     "createTable": " node -r @babel/register server/migrations/createTable.js",
     "createTable:test": "cross-env NODE_ENV=test babel-node server/migrations/createTable.js",
     "accountTable": "node -r @babel/register server/migrations/account.js",
@@ -49,7 +49,6 @@
     "coveralls": "^3.0.3",
     "cross-env": "^5.2.0",
     "dotenv": "^7.0.0",
-    "eslint": "^5.16.0",
     "express": "^4.16.4",
     "express-validator": "^5.3.1",
     "jol": "^0.1.0",
@@ -64,6 +63,7 @@
   },
   "devDependencies": {
     "@babel/cli": "^7.4.3",
+    "eslint": "^5.16.0",
     "mocha-lcov-reporter": "^1.3.0",
     "nyc": "^13.3.0"
   }
