+++ server/Controllers/UserController.js
@@ -52,7 +52,7 @@ class UserController{
             };
             const userResult = await db.query(sql);
   
-              jwt.sign({ user: userResult.rows[0].id,admin:userResult.rows[0].admin}, process.env.SECRET_KEY, (err, token) => 
+              jwt.sign({ user: userResult.rows[0].id,admin:userResult.rows[0].admin, cashier: userResult.rows[0]}, process.env.SECRET_KEY, (err, token) => 
                res.status(201).json({
                 success: true,
                 status: '201',
@@ -100,7 +100,7 @@ static async login (req, res) {
                 if (result && result.rows.length === 1) {
                   delete result.rows[0].password;
                   
-                  jwt.sign({ user: result.rows[0].id, admin:result.rows[0].admin}, process.env.SECRET_KEY, (err, token) =>
+                  jwt.sign({ user: result.rows[0].id, admin:result.rows[0].admin, cashier:result.rows[0].type}, process.env.SECRET_KEY, (err, token) =>
                     res.status(201).json({
                     success: true,
                     message: 'user successful login',
