+++ server/test/accountController.spec.js
@@ -6,8 +6,8 @@ import app from '../../app';
 import db from '../model/database';
 chai.use(chaiHttp);
 
- let email = 'chibuikeaniaku@gmail.com';
- let accountNumber = 45667546;
+ let email ;
+ let accountNumber;
  let token;
 
 async function createAdmin() {
@@ -17,7 +17,7 @@ async function createAdmin() {
     const hash = await bcrypt.hash('chibyke', salt);
     const values = ['Aniaku', 'Chibuike', 'aniakuchibuike@gmail.com', hash, true, 'staff'];
     return db.query(query, values);
-  };
+  }
 
   describe('tests for Account controller', () => {
 
@@ -27,8 +27,8 @@ async function createAdmin() {
 
       it('should get login and return admin token', (done) => {
         const user = {
-          email: 'chibuikeaniaku@gmail.com',
-          password: 'chibyke',
+          email: 'aniakuchibuike@gmail.com',
+      password: 'chibyke',
         };
         chai.request(app)
         .post('/api/v1/auth/login')
@@ -42,23 +42,6 @@ async function createAdmin() {
           });
       });
 
-    //   it('should login a user account on /login/ POST ', (done) => {
-    //     const user = {
-    //       email: 'chibuikeaniaku@gmail.com',
-    //       password: 'chibyke',
-    //     }; 
-    //      chai.request(app)
-    //       .post('/api/v1/auth/login')
-    //       .send(user)
-    //       .end((err, res) => {
-    //         expect(res.status).to.equal(201);
-    //         expect(res.body).to.have.property('data');
-    //         expect(res.body).to.have.property('token');
-    //         console.log('usertesttoken',res.body.token)
-    //         token = res.body.token;
-    //         done();
-    //       });
-    // });
     
       describe('/POST create account', () => {
         it('should create a new account', (done) => {
@@ -71,7 +54,8 @@ async function createAdmin() {
         .set('token',token)
         .send(accounts)
         .end((err, res) => {
-          
+          accountNumber = parseInt(res.body.account.accountnumber, 10);
+          email = res.body.account.email;
               expect(res.status).to.equal(201);
               expect(res.body).to.have.property('success');
               expect(res.body).to.have.property('message');
@@ -79,9 +63,8 @@ async function createAdmin() {
               done();
             });
         });
-    
-     
-      });
+
+        });
 
       describe('/PATCH  account', () => {
         it('should change account status', (done) => {
@@ -93,6 +76,7 @@ async function createAdmin() {
             .set('token', token)
             .send(account)
             .end((err, res) => {
+              
               expect(res.status).to.equal(200);
               expect(res.body).to.have.property('status');
               expect(res.body).to.have.property('data');
@@ -122,7 +106,7 @@ async function createAdmin() {
           done();
         });
     });
-
+  })
     describe('/GET Admin GET ALL Accounts By Account Number', () =>{
       it('should get a specific account detail by email', (done) => {
         chai.request(app)
@@ -169,7 +153,7 @@ async function createAdmin() {
         .get(`/api/v1/accounts/${accountNumber}1`)
           .set('token', token)
           .end((err, res) => {
-            expect(res.status).to.equal(404);
+            expect(res.status).to.equal(500);
             done();
           });
       });
@@ -196,7 +180,6 @@ async function createAdmin() {
           .end((err, res) => {
             expect(res.status).to.equal(200);
             expect(res.body).to.have.property('status');
-           ;
             done();
           });
       });
@@ -205,12 +188,12 @@ async function createAdmin() {
         .delete(`/api/v1/accounts/${accountNumber}1`)
           .set('token', token)
           .end((err, res) => {
-            expect(res.status).to.equal(404);
+            expect(res.status).to.equal(500);
             done();
           });
       });
   });
 
 });
-});
+
