+++ swagger.yaml
@@ -1,0 +1,13 @@
+swagger: "2.0"
+info: 
+  version: "1.08"
+  host: localhost:5000
+  basePath: /
+  schemes:
+    -http
+  consumes:
+    - application/json
+  produces:
+    - application/json
+  paths:
+    - C:\Users\ONews2\Documents\My-Banka\swagger.yaml
